+++ BidmadPluginSample/Assets/Plugins/iOS/Bidmad/BidmadSDK.framework/Headers/BidmadSDK.h
@@ -19,13 +19,11 @@ FOUNDATION_EXPORT const unsigned char BidmadSDKVersionString[];
 #import "BIDMADSetting.h"
 #import "BIDMADBanner.h"
 #import "BIDMADInterstitial.h"
-
 #import "BIDMADRewardVideo.h"
 #import "BIDMADAppOpenAd.h"
-#import "BIDMADFacebook.h"
-#import "BIDMADFacebookBidding.h"
-#import "BIDMADAdmob.h"
 
+#import "BIDMADFacebook.h"
+#import "BIDMADAdmob.h"
 #import "BIDMADAdmanager.h"
 #import "BIDMADAtom.h"
 #import "BIDMADAtomInterstitial.h"
@@ -33,3 +31,6 @@ FOUNDATION_EXPORT const unsigned char BidmadSDKVersionString[];
 #import "BIDMADLabelView.h"
 #import "UIButton+circle.h"
 
+#import "UnrealCommon.h"
+#import "UnrealInterstitial.h"
+#import "UnrealReward.h"